refactor: all providers should be centralized one place
