+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
-import { CounterStoreProvider } from '@/provider/counterProvider'
-import { QueryProvider } from '@/provider/queryProvider'
+import { Providers } from '@/providers'
 
 import './globals.css'
 
@@ -28,9 +27,7 @@ export default function RootLayout({
   return (
     <html lang='en'>
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <QueryProvider>
-          <CounterStoreProvider>{children}</CounterStoreProvider>
-        </QueryProvider>
+        <Providers>{children}</Providers>
       </body>
     </html>
   )
